requirements.txt
alembic==0.8.8
blinker==1.4
click==6.6
Flask==0.11.1
Flask-Mail==0.9.1
Flask-Mako==0.4
Flask-Migrate==2.0.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.4
MarkupSafe==0.23
psycopg2==2.6.2
PyHAML==1.1.0
python-editor==1.0.1
six==1.10.0
SQLAlchemy==1.0.15
Werkzeug==0.11.11
WTForms==2.1
